Big revision: interface and Arduino communication
Integration of different modes to same interface screen.

Changelog:
- (Warning: buttons for realtime classification were deactivated, code
is also not finished.)
- New interface layout
- Created fNIRSData class to store and handle the processing of fNIRS
data
- fNIRS.m is no longer used, integrated into fNIRSData class
- ReceiveDataFromArduino is no longer used, integrated into GUI_fNIRS.m
- Processing of data is only started after user interaction.
Pre-processing is done automatically after the acquisition of data. This
determines if enough samples are captured to run the necessary bandpass
filtering used in both the decimation and processing phase. The data is
also automatically downsampled with the a user input factor if the
previous check has been passed.
- Error handling and status changes: check user input, disable/enable
buttons, change text on buttons, output to status message field.
- Rewritten Arduino communication to process commands from and to Matlab
better. The matlab serial read qeue is flushed before new data is
received. The Arduino program was restructured because some commands
were not processed correctly.